Time: 0 ms (100%), Space: 8.5 MB (54.61%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int reverse(int x) {
 
-     long long int revNum=0;
+        long int revNum=0;
 
         while(x!=0){
             revNum=(revNum*10)+(x%10);
